+++ PointOfSale/PointOfSale/Terminal/PriceCalculations/ComplexPriceCalculator.cs
@@ -10,20 +10,20 @@ namespace PointOfSale.Terminal.PriceCalculations
     internal class ComplexPriceCalculator : IPriceCalculator
     {
         private readonly SingleUnitPriceCalculator singleUnitCalculator;
-        private readonly double volumePrice;
+        private readonly decimal volumePrice;
         private readonly int volumeSize;
 
-        public ComplexPriceCalculator(double singleUnitPrice, int volumeSize, double volumePrice)
+        public ComplexPriceCalculator(decimal singleUnitPrice, int volumeSize, decimal volumePrice)
         {
             Contract.Requires(volumeSize > 0);
-            Contract.Requires(volumePrice > 0.0);
+            Contract.Requires(volumePrice > 0.0M);
 
             singleUnitCalculator = new SingleUnitPriceCalculator(singleUnitPrice);
             this.volumeSize = volumeSize;
             this.volumePrice = volumePrice;
         }
 
-        public double CalculatePrice(int itemsCount, double discountRate)
+        public decimal CalculatePrice(int itemsCount, decimal discountRate)
         {
             return (itemsCount / volumeSize) * volumePrice + singleUnitCalculator.CalculatePrice(itemsCount % volumeSize, discountRate);
         }
